// apps/server/src/generated/prisma/models/CrmOrganizationConfig.ts
export type CrmOrganizationConfigModel = runtime.Types.Result.DefaultSelection<Prisma.$CrmOrganizationConfigPayload>;
// ...
export type $CrmOrganizationConfigPayload<
  ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs
> = {
  name: 'CrmOrganizationConfig';
  objects: {
    organization: Prisma.$OrganizationPayload<ExtArgs>;
  };
  scalars: runtime.Types.Extensions.GetPayloadResult<
    {
      id: string;
      organizationId: string;
      allowAdminViewMemberEmailBody: boolean;
      updatedById: string | null;
      updatedByName: string | null;
      createdAt: Date;
      updatedAt: Date;
    },
    ExtArgs['result']['crmOrganizationConfig']
  >;
  composites: {};
};